Test the FileSet class instead of just the pattern.
The original tests did not actually test any of the source code. Refactored the partition class to store the PARTITIONING_PATTERN as a module-level attribute that can be imported by other modules. Rewrote the test of the pattern to be a test of the partition_by() method.

Added an __init__.py to make tests/partition into a Python package.